+++ addons/godotdevconsole/panels/terminal/scripts/commands/CreatePanelCMD.cs
@@ -1,4 +1,3 @@
-
 namespace GodotDevConsole.Panels.Terminal.Commands
 {
     public static class CreatePanelCMD
@@ -6,7 +5,17 @@ namespace GodotDevConsole.Panels.Terminal.Commands
         [Command(Aliases = new string[] { "mkpan" }, Description = "Creates a panel of a given type.")]
         public static void CreatePanel(string panelName, string panelTypeName, TerminalPanel panel)
         {
-            DevConsole.Instance.CreatePanel(panelName, panelTypeName);
+            switch (DevConsole.Instance.CreatePanel(panelName, panelTypeName))
+            {
+                case CreatePanelMessage.Success:
+                    break;
+                case CreatePanelMessage.InvalidPanelType:
+                    panel.Print($"{panelTypeName} is not a valid panel type.");
+                    break;
+                case CreatePanelMessage.PanelWithSameName:
+                    panel.Print($"A panel with name {panelName} already exists.");
+                    break;
+            }
         }
     }
 }