+++ addons/godotdevconsole/panels/terminal/scripts/commands/DestroyPanelCMD.cs
@@ -1,4 +1,3 @@
-
 namespace GodotDevConsole.Panels.Terminal.Commands
 {
     public static class DestroyPanelCMD
@@ -6,7 +5,18 @@ namespace GodotDevConsole.Panels.Terminal.Commands
         [Command(Aliases = new string[] { "rmpan" }, Description = "Removes a panel by name.")]
         public static void DestroyPanel(string panelName, TerminalPanel panel)
         {
-            panel.Print(DevConsole.Instance.DestroyPanel(panelName));
+            switch (DevConsole.Instance.DestroyPanel(panelName))
+            {
+                case DestroyPanelMessage.Success:
+                    panel.Print($"Panel {panelName} destroyed");
+                    break;
+                case DestroyPanelMessage.NoPanelWithName:
+                    panel.Print($"Could not find panel with name {panelName}");
+                    break;
+                case DestroyPanelMessage.LastPanel:
+                    panel.Print("Can't destroy last panel.");
+                    break;
+            }
         }
     }
 }